Refactor code and remove unused UI components
