+++ src/main/java/duke/Launcher.java
@@ -7,6 +7,6 @@ import static javafx.application.Application.launch;
  */
 public class Launcher {
     public static void main(String[] args) {
-        launch(Main.class, args);
+        launch(Duke.class, args);
     }
 }
