--- src/main/java/duke/Main.java
@@ -1,30 +1,0 @@
-package duke;
-
-import java.io.IOException;
-
-import duke.gui.MainWindow;
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.layout.AnchorPane;
-import javafx.stage.Stage;
-
-/**
- * A GUI for Duke using FXML.
- */
-public class Main extends Application {
-    @Override
-    public void start(Stage stage) {
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
-            AnchorPane ap = fxmlLoader.load();
-            Scene scene = new Scene(ap);
-            stage.setScene(scene);
-            stage.setResizable(false);
-            stage.show();
-            fxmlLoader.<MainWindow>getController().setDuke(new Duke());
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-}
