+++ src/main/java/duke/command/ExitCommand.java
@@ -2,8 +2,6 @@ package duke.command;
 
 import duke.task.TaskList;
 
-import static duke.common.Messages.GOODBYE_MSG;
-
 class ExitCommand implements Command {
     @Override
     public boolean isExitCommand() {
@@ -12,6 +10,6 @@ class ExitCommand implements Command {
 
     @Override
     public CommandResponse execute(TaskList taskList) {
-        return new CommandResponse(GOODBYE_MSG);
+        return new CommandResponse();
     }
 }
