+++ src/main/java/duke/ui/DialogBox.java
@@ -1,4 +1,4 @@
-package duke.gui;
+package duke.ui;
 
 import java.io.IOException;
 import java.util.Collections;
